reshuffle dtos & models (no logic change)
