+++ api/handle_scenarios.go
@@ -2,44 +2,12 @@ package api
 
 import (
 	"net/http"
-	"time"
 
 	"github.com/gin-gonic/gin"
 
 	"github.com/checkmarble/marble-backend/dto"
-	"github.com/checkmarble/marble-backend/models"
 	"github.com/checkmarble/marble-backend/pure_utils"
 )
-
-type APIScenario struct {
-	Id                     string    `json:"id"`
-	CreatedAt              time.Time `json:"createdAt"`
-	DecisionToCaseOutcomes []string  `json:"decision_to_case_outcomes"`
-	DecisionToCaseInboxId  string    `json:"decision_to_case_inbox_id"`
-	Description            string    `json:"description"`
-	LiveVersionID          *string   `json:"liveVersionId,omitempty"`
-	Name                   string    `json:"name"`
-	OrganizationId         string    `json:"organization_id"`
-	TriggerObjectType      string    `json:"triggerObjectType"`
-}
-
-func NewAPIScenario(scenario models.Scenario) APIScenario {
-	out := APIScenario{
-		Id:        scenario.Id,
-		CreatedAt: scenario.CreatedAt,
-		DecisionToCaseOutcomes: pure_utils.Map(scenario.DecisionToCaseOutcomes,
-			func(o models.Outcome) string { return o.String() }),
-		Description:       scenario.Description,
-		LiveVersionID:     scenario.LiveVersionID,
-		Name:              scenario.Name,
-		OrganizationId:    scenario.OrganizationId,
-		TriggerObjectType: scenario.TriggerObjectType,
-	}
-	if scenario.DecisionToCaseInboxId != nil {
-		out.DecisionToCaseInboxId = *scenario.DecisionToCaseInboxId
-	}
-	return out
-}
 
 func (api *API) ListScenarios(c *gin.Context) {
 	usecase := api.UsecasesWithCreds(c.Request).NewScenarioUsecase()
@@ -47,7 +15,7 @@ func (api *API) ListScenarios(c *gin.Context) {
 	if presentError(c, err) {
 		return
 	}
-	c.JSON(http.StatusOK, pure_utils.Map(scenarios, NewAPIScenario))
+	c.JSON(http.StatusOK, pure_utils.Map(scenarios, dto.AdaptScenarioDto))
 }
 
 func (api *API) CreateScenario(c *gin.Context) {
@@ -58,11 +26,11 @@ func (api *API) CreateScenario(c *gin.Context) {
 	}
 
 	usecase := api.UsecasesWithCreds(c.Request).NewScenarioUsecase()
-	scenario, err := usecase.CreateScenario(c.Request.Context(), dto.AdaptCreateScenario(input))
+	scenario, err := usecase.CreateScenario(c.Request.Context(), dto.AdaptCreateScenarioInput(input))
 	if presentError(c, err) {
 		return
 	}
-	c.JSON(http.StatusOK, NewAPIScenario(scenario))
+	c.JSON(http.StatusOK, dto.AdaptScenarioDto(scenario))
 }
 
 func (api *API) GetScenario(c *gin.Context) {
@@ -74,7 +42,7 @@ func (api *API) GetScenario(c *gin.Context) {
 	if presentError(c, err) {
 		return
 	}
-	c.JSON(http.StatusOK, NewAPIScenario(scenario))
+	c.JSON(http.StatusOK, dto.AdaptScenarioDto(scenario))
 }
 
 func (api *API) UpdateScenario(c *gin.Context) {
@@ -89,9 +57,9 @@ func (api *API) UpdateScenario(c *gin.Context) {
 
 	scenario, err := usecase.UpdateScenario(
 		c.Request.Context(),
-		dto.AdaptUpdateScenario(scenarioId, input))
+		dto.AdaptUpdateScenarioInput(scenarioId, input))
 	if presentError(c, err) {
 		return
 	}
-	c.JSON(http.StatusOK, NewAPIScenario(scenario))
+	c.JSON(http.StatusOK, dto.AdaptScenarioDto(scenario))
 }
