+++ dto/scenarios.go
@@ -1,21 +1,60 @@
 package dto
 
 import (
+	"time"
+
 	"github.com/checkmarble/marble-backend/models"
 	"github.com/checkmarble/marble-backend/pure_utils"
 	"github.com/guregu/null/v5"
 )
 
+// Read DTO
+type ScenarioDto struct {
+	Id                     string    `json:"id"`
+	CreatedAt              time.Time `json:"createdAt"`
+	DecisionToCaseOutcomes []string  `json:"decision_to_case_outcomes"`
+	DecisionToCaseInboxId  string    `json:"decision_to_case_inbox_id"`
+	Description            string    `json:"description"`
+	LiveVersionID          *string   `json:"liveVersionId,omitempty"`
+	Name                   string    `json:"name"`
+	OrganizationId         string    `json:"organization_id"`
+	TriggerObjectType      string    `json:"triggerObjectType"`
+}
+
+func AdaptScenarioDto(scenario models.Scenario) ScenarioDto {
+	out := ScenarioDto{
+		Id:        scenario.Id,
+		CreatedAt: scenario.CreatedAt,
+		DecisionToCaseOutcomes: pure_utils.Map(scenario.DecisionToCaseOutcomes,
+			func(o models.Outcome) string { return o.String() }),
+		Description:       scenario.Description,
+		LiveVersionID:     scenario.LiveVersionID,
+		Name:              scenario.Name,
+		OrganizationId:    scenario.OrganizationId,
+		TriggerObjectType: scenario.TriggerObjectType,
+	}
+	if scenario.DecisionToCaseInboxId != nil {
+		out.DecisionToCaseInboxId = *scenario.DecisionToCaseInboxId
+	}
+	return out
+}
+
+// Create scenario DTO
 type CreateScenarioBody struct {
 	Name              string `json:"name"`
 	Description       string `json:"description"`
 	TriggerObjectType string `json:"triggerObjectType"`
 }
 
-type CreateScenarioInput struct {
-	Body *CreateScenarioBody `in:"body=json"`
+func AdaptCreateScenarioInput(input CreateScenarioBody) models.CreateScenarioInput {
+	return models.CreateScenarioInput{
+		Name:              input.Name,
+		Description:       input.Description,
+		TriggerObjectType: input.TriggerObjectType,
+	}
 }
 
+// Update scenario DTO
 type UpdateScenarioBody struct {
 	DecisionToCaseOutcomes []string    `json:"decision_to_case_outcomes"`
 	DecisionToCaseInboxId  null.String `json:"decision_to_case_inbox_id"`
@@ -23,7 +62,7 @@ type UpdateScenarioBody struct {
 	Name                   *string     `json:"name"`
 }
 
-func AdaptUpdateScenario(scenarioId string, input UpdateScenarioBody) models.UpdateScenarioInput {
+func AdaptUpdateScenarioInput(scenarioId string, input UpdateScenarioBody) models.UpdateScenarioInput {
 	parsedInput := models.UpdateScenarioInput{
 		Id:                    scenarioId,
 		DecisionToCaseInboxId: input.DecisionToCaseInboxId,
@@ -34,45 +73,4 @@ func AdaptUpdateScenario(scenarioId string, input UpdateScenarioBody) models.Upd
 		parsedInput.DecisionToCaseOutcomes = pure_utils.Map(input.DecisionToCaseOutcomes, models.OutcomeFrom)
 	}
 	return parsedInput
-}
-
-// Scenario iterations
-type UpdateScenarioIterationData struct {
-	TriggerConditionAstExpression *NodeDto `json:"trigger_condition_ast_expression"`
-	ScoreReviewThreshold          *int     `json:"scoreReviewThreshold,omitempty"`
-	ScoreRejectThreshold          *int     `json:"scoreRejectThreshold,omitempty"`
-	Schedule                      *string  `json:"schedule"`
-	BatchTriggerSQL               *string  `json:"batchTriggerSQL"`
-}
-
-type UpdateScenarioIterationBody struct {
-	Body *UpdateScenarioIterationData `json:"body,omitempty"`
-}
-
-type CreateScenarioIterationBody struct {
-	ScenarioId string `json:"scenarioId"`
-	Body       *struct {
-		TriggerConditionAstExpression *NodeDto              `json:"trigger_condition_ast_expression"`
-		Rules                         []CreateRuleInputBody `json:"rules"`
-		ScoreReviewThreshold          *int                  `json:"scoreReviewThreshold,omitempty"`
-		ScoreRejectThreshold          *int                  `json:"scoreRejectThreshold,omitempty"`
-		Schedule                      string                `json:"schedule"`
-		BatchTriggerSQL               string                `json:"batchTriggerSQL"`
-	} `json:"body,omitempty"`
-}
-
-type CreateScenarioIterationInput struct {
-	Payload *CreateScenarioIterationBody `in:"body=json"`
-}
-
-type CreateDraftFromScenarioIterationInput struct {
-	ScenarioIterationId string `in:"path=scenarioIterationId"`
-}
-
-func AdaptCreateScenario(input CreateScenarioBody) models.CreateScenarioInput {
-	return models.CreateScenarioInput{
-		Name:              input.Name,
-		Description:       input.Description,
-		TriggerObjectType: input.TriggerObjectType,
-	}
 }
